requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
scikit-learn
datasets
sentence-transformers
faiss-cpu
langgraph
langchain
python-dotenv
